+++ map_utils/lib_copy.go
@@ -2,6 +2,9 @@
 package map_utils
 
 // CopyMap takes a map of strings to interfaces and returns a copy of that map
+// with all the values copied as well.
+// - m: the map to copy
+// - returns: a copy of the map
 func CopyMap(m map[string]interface{}) map[string]interface{} {
 	cp := make(map[string]interface{})
 	for k, v := range m {
