+++ math/lib_boolean.go
@@ -2,16 +2,24 @@
 package math
 
 // The NOT function returns the opposite of the boolean value passed to it.
+// - a: the boolean value to negate
+// - returns: the opposite of the boolean value passed to it
 func NOT(a bool) bool {
 	return !a
 }
 
-// `OR` returns `true` if either `a` or `b` is `true`, otherwise it returns `false`
+// `OR` the logical	or of its two boolean arguments
+// - a: the first boolean value
+// - b: the second boolean value
+// - returns: `true` if either `a` or `b` is `true`, otherwise it returns `false`
 func OR(a bool, b bool) bool {
 	return a || b
 }
 
-// `AND` returns the logical AND of its two boolean arguments
+// `AND` the logical AND of its two boolean arguments
+// - a: the first boolean value
+// - b: the second boolean value
+// - returns: `true` if both `a` and `b` are `true`, otherwise it returns `false`
 func AND(a bool, b bool) bool {
 	return a && b
 }
