+++ math/lib_math.go
@@ -1,21 +1,33 @@
 package math
 
 // Minus subtracts two integers
+// - a: the first integer
+// - b: the second integer
+// - returns: the difference of the two integers
 func Minus(a, b int) int {
 	return a - b
 }
 
 // Plus adds two integers
+// - a: the first integer
+// - b: the second integer
+// - returns: the sum of the two integers
 func Plus(a, b int) int {
 	return a + b
 }
 
 // Multiply multiplies two integers
+// - a: the first integer
+// - b: the second integer
+// - returns: the product of the two integers
 func Multiply(a, b int) int {
 	return a * b
 }
 
 // Divide divides two integers
+// - a: the first integer
+// - b: the second integer
+// - returns: the quotient of the two integers
 func Divide(a, b int) int {
 	return a / b
 }
