+++ net/lib_api.go
@@ -20,6 +20,9 @@ type Header struct {
 
 // It takes a username and password, concatenates them with a colon, encodes the resulting string in
 // base64, and returns the encoded string
+// - username: the username
+// - password: the password
+// - returns: the encoded string
 func BasicAuth(username, password string) string {
 	auth := username + ":" + password
 	return base64.StdEncoding.EncodeToString([]byte(auth))
@@ -27,6 +30,12 @@ func BasicAuth(username, password string) string {
 
 // CallApi() is a function that takes a URL, a method, a list of headers, a payload, and a boolean to
 // ignore SSL errors. It returns a byte array of result data, a boolean, and an error
+// - url: the URL to call
+// - method: the HTTP method to use POST GET PATH etc
+// - headers: a list of headers to send with the request
+// - payload: the data to send with the request
+// - ignore_ssl: a boolean to ignore SSL errors
+// - returns: a byte array of result data, a boolean is success, and an error if there is one
 func CallApi(url string, method string, headers []Header, payload io.Reader, ignore_ssl bool) ([]byte, bool, error) {
 	//****************************
 	//Create the connection string
@@ -65,7 +74,9 @@ func CallApi(url string, method string, headers []Header, payload io.Reader, ign
 	return responseData, true, nil
 }
 
-// If the host string contains either "http://" or "https://", return nil. Otherwise, return an error
+// CheckValidUrl Check If the host string contains either "http://" or "https://", return nil. Otherwise, return an error
+// - host: the host string to check
+// - returns: an error if the host string does not contain either "http://" or "https://"
 func CheckValidUrl(host string) error {
 	//Validate is a URL
 	if strings.Contains(host, "http://") || strings.Contains(host, "https://") {
@@ -75,6 +86,8 @@ func CheckValidUrl(host string) error {
 }
 
 // RemoveURL removes http:// or https:// from a string
+// - str: the string to remove http:// or https:// from
+// - returns: the string without http:// or https://
 func RemoveURL(str string) string {
 	str = strings.ReplaceAll(str, "https://", "")
 	str = strings.ReplaceAll(str, "http://", "")
